+++ MuseumRewardsIn/MuseumRewardsIn/docs/Changelog.md
@@ -3,7 +3,8 @@
 
 #### Version 0.2.0
 
-Adjustments for SVE. Port to using AtraCore.
+* Adjustments for SVE. Port to using AtraCore.
+* Fix Dwarven Translation Guide showing up in the store.
 
 #### Version 0.1.0